backend/requirements.txt
faiss-cpu
sentence-transformers
fastapi
uvicorn
anthropic
google-api-python-client 
google-auth 
google-auth-oauthlib 
python-multipart 
pypdf 
python-docx 
pandas 
openpyxl
celery 
redis
requests
python-dotenv
yfinance==0.2.40
sec-edgar-downloader==5.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
